requirements.txt
gradio
requests
python-dotenv
